Replace "div" with "span" inside "p"
The permitted contents of a paragraph are only phrasing contents, so
"div" elements can not be used, although "span" can. Besides being
invalid HTML it seems that the browser ends the paragraph at the div, so
the label/input that appears after it is treated as being in a new
paragraph, which is not indented by default, and thus is not aligned
with the rest of its sibling inputs.

Note that an additional div is nevertheless added once the page is
loaded to be able to select the groups, but this one does not break its
parent paragraph (maybe due to being added after the page load, but I do
not really know). Nevertheless, it needs to be explicitly indented, and
the second indentend wrapper needs to be removed, as it affects only the
label but not the div/input, and therefore the label had an extra
indentation over the input.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>