Disable "Enforce password protection" if not asking for a password
"Enforce password protection" is a subcase of "Always ask for a
password", so it should be disabled if its parent option is unchecked.
Although other dependant options in the sharing settings are fully
hidden instead of just disabled this option is disabled but shown to
avoid confusion when updating from a previous Nextcloud version where
"Always ask for a password" was unchecked and "Enforce password
protection" was checked.

Besides that, due to their dependency the enforced password protection
is now automatically unchecked too if its parent option is unchecked.

Finally, the groups excluded from password requirements are also
disabled when "Always ask for a password" is unchecked, as they might be
still shown if "Enforce password protection" was checked due to an
update from a previous version.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>